create requests helper
