+++ notifiers/__init__.py
@@ -3,4 +3,4 @@ import logging
 from ._version import __version__
 from .core import get_notifier, all_providers
 
-logging.getLogger('notifiers').addHandler(logging.NullHandler())
+logger = logging.getLogger('notifiers').addHandler(logging.NullHandler())
